+++ server/src/main/java/com/project/game/player/BaseStats.java
@@ -2,17 +2,17 @@ package com.project.game.player;
 
 /** Durable/progression source statistics for a player. */
 public record BaseStats(
-        long hp,
-        long mp,
-        long damage,
-        long armor,
+        int hp,
+        int mp,
+        int damage,
+        int armor,
         int critical,
         int dodge,
-        long constitution,
+        int constitution,
         int speed) {
     public BaseStats {
-        if (hp < 0L || mp < 0L || damage < 0L || armor < 0L
-                || critical < 0 || dodge < 0 || constitution < 0L || speed <= 0) {
+        if (hp < 0 || mp < 0 || damage < 0 || armor < 0
+                || critical < 0 || dodge < 0 || constitution < 0 || speed <= 0) {
             throw new IllegalArgumentException("base stats must be non-negative and speed must be positive");
         }
     }
