+++ server/src/main/java/com/project/game/player/CurrentStats.java
@@ -2,17 +2,17 @@ package com.project.game.player;
 
 /** Durable snapshot of effective player statistics used by realtime combat. */
 public record CurrentStats(
-        long maxHp,
-        long maxMp,
-        long damage,
-        long armor,
+        int maxHp,
+        int maxMp,
+        int damage,
+        int armor,
         int critical,
         int dodge,
-        long constitution,
+        int constitution,
         int speed) {
     public CurrentStats {
-        if (maxHp < 0L || maxMp < 0L || damage < 0L || armor < 0L
-                || critical < 0 || dodge < 0 || constitution < 0L || speed <= 0) {
+        if (maxHp < 0 || maxMp < 0 || damage < 0 || armor < 0
+                || critical < 0 || dodge < 0 || constitution < 0 || speed <= 0) {
             throw new IllegalArgumentException("current stats must be non-negative and speed must be positive");
         }
     }
